Fix some lint errors

+++ crates/sort/benches/osort.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::collection_is_never_read)]
 #![allow(missing_docs)]
 use criterion::{
   criterion_group, criterion_main, measurement::Measurement, AxisScale, BenchmarkId, Criterion,
@@ -11,13 +12,13 @@ use std::hint::black_box;
 
 pub fn benchmark_sort<T: Measurement + 'static>(c: &mut Criterion<T>) {
   let mut group =
-    c.benchmark_group(format!("Sorting/{}", std::any::type_name::<T>().split(":").last().unwrap()));
+    c.benchmark_group(format!("Sorting/{}", std::any::type_name::<T>().split(':').last().unwrap()));
   let plot_config = PlotConfiguration::default().summary_scale(AxisScale::Logarithmic);
   group.plot_config(plot_config);
 
   for &size in &[100, 1_000, 3_162, 10_000, 31_623, 100_000] {
     group.bench_with_input(BenchmarkId::new("Bitonic", size), &size, |b, &size| {
-      let mut data: Vec<i32> = (0..size as i32).collect();
+      let mut data: Vec<i32> = (0..size).collect();
       data.shuffle(&mut rand::thread_rng());
       let data = data;
       b.iter(|| {
@@ -27,7 +28,7 @@ pub fn benchmark_sort<T: Measurement + 'static>(c: &mut Criterion<T>) {
     });
 
     group.bench_with_input(BenchmarkId::new("Batcher", size), &size, |b, &size| {
-      let mut data: Vec<i32> = (0..size as i32).collect();
+      let mut data: Vec<i32> = (0..size).collect();
       data.shuffle(&mut rand::thread_rng());
       let data = data;
       b.iter(|| {
@@ -38,7 +39,7 @@ pub fn benchmark_sort<T: Measurement + 'static>(c: &mut Criterion<T>) {
     });
 
     group.bench_with_input(BenchmarkId::new("std::sort", size), &size, |b, &size| {
-      let mut data: Vec<i32> = (0..size as i32).collect();
+      let mut data: Vec<i32> = (0..size).collect();
       data.shuffle(&mut rand::thread_rng());
       let data = data;
       b.iter(|| {
